+++ app/controllers/better_mailer_previews/application_controller.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module BetterMailerPreviews
-  class ApplicationController < ActionController::Base
-  end
-end