+++ app/jobs/better_mailer_previews/application_job.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module BetterMailerPreviews
-  class ApplicationJob < ActiveJob::Base
-  end
-end