+++ spec/dummy/spec/mailers/previews/daily_preview_mailer_preview.rb
@@ -1,0 +1,5 @@
+class DailyPreviewMailerPreview < ActionMailer::Preview
+  def digest
+    DailyPreviewMailer.digest
+  end
+end